another converson from double to int
svn-id: https://kwave.svn.sourceforge.net/svnroot/kwave/trunk@803 bbf32e5a-961d-0410-9e10-91f32741c31e

+++ plugins/selectrange/SelectRangeDialog.cpp
@@ -124,7 +124,7 @@ SelectRangeDialog::SelectRangeDialog(QWidget *widget,
 	    samplesChanged(0);
 	    break;
 	case SelectRangePlugin::byPercents:
-	    percentsChanged(m_range);
+	    percentsChanged((int)m_range);
 	    break;
     }
